size of variable

+++ 0x00-hello_world/6-size.c
@@ -13,10 +13,10 @@ int main(void)
 	long long int D;
 	float F;
 
-	printf("Size of a char: %zu byte(s)\n", sizeof(A));
-	printf("Size of an int: %zu byte(s)\n", sizeof(B));
-	printf("Size of a long int: %zu byte(s)\n", sizeof(C));
-	printf("Size of a long long int: %zu byte(s)\n", sizeof(D);
-	printf("Size of a float: %zu byte(s)\n", sizeof(F);
+	printf("Size of a char: %zu byte(s)\n", (unsigned long)sizeof(A));
+	printf("Size of an int: %zu byte(s)\n", (unsigned long)sizeof(B));
+	printf("Size of a long int: %zu byte(s)\n", (unsigned long)sizeof(C));
+	printf("Size of a long long int: %zu byte(s)\n", (unsigned long)sizeof(D);
+	printf("Size of a float: %zu byte(s)\n", (unsigned long)sizeof(F);
 	return (0);
 }
